feat: add create menu event

+++ BuberDinner.Domain/Common/ValueObjects/AverageRating.cs
@@ -4,17 +4,17 @@ namespace BuberDinner.Domain.Common.ValueObjects
 {
     public sealed class AverageRating : ValueObject
     {
-        public double Value { get; private set; }
-        public double NumRatings { get; private set; }
+        public float Value { get; private set; }
+        public int NumRatings { get; private set; }
 
-        private AverageRating(double value, int numRatings)
+        private AverageRating(float value, int numRatings)
         {
             Value = value;
             NumRatings = numRatings;
         }
 
         public static AverageRating CreateNew(
-            double value = 0,
+            float value = 0,
             int numRatings = 0)
         {
             return new AverageRating(value, numRatings);
